Changed type conversion of input numbers
Changed type conversion of input numbers from int to float to handle floating numbers

+++ sum.py
@@ -1,10 +1,10 @@
 # This program gets two numbers from user, sums them up and displays the sum value
 # Get first number from User
-num1 = int( input('Enter first number: ') )
+num1 = float( input('Enter first number: ') )
 print(num1)
 
 # Get second number from User
-num2 = int( input('Enter second number: ') )
+num2 = float( input('Enter second number: ') )
 print(num2)
 
 # Calculate sum of the two numbers and print the sum
